Fix compile warnings.

+++ tests/loadtests/vkloadtests/Texture.cpp
@@ -587,7 +587,7 @@ Texture::prepareSamplerAndView()
     samplerInfo.magFilter = vk::Filter::eLinear;
     samplerInfo.minFilter = vk::Filter::eLinear;
     samplerInfo.mipmapMode = vk::SamplerMipmapMode::eLinear;
-    samplerInfo.maxLod = texture.levelCount;
+    samplerInfo.maxLod = (float)texture.levelCount;
     samplerInfo.anisotropyEnable = true;
     samplerInfo.maxAnisotropy = 8;
     samplerInfo.borderColor = vk::BorderColor::eFloatOpaqueWhite;
